Add files via upload
In this task, I write one of the basic Python tools - for loop
What to do: 
All too-long words (more than 10 letters) can be replaced with a special abbreviation.
This abbreviation is built as follows: the first and last letter of the word is written, and between them - the number of letters between the first and last letter (in decimal notation and without leading zeros). Thus, "destination" would be written as "d9n".